KVM: vmx: add basic handling of VMEXIT from SGX enclave
Define the new bits added to VM_EXIT_REASON and GUEST_INTERRUPT_STATE
to denote a VMEXIT from a SGX enclave.  Cache and clear the enclave
bit in VM_EXIT_REASON immediately after VMEXIT so that we do not have
to modify existing logic that assumes no modifier bits are set in the
cache exit_reason variable.

Add a WARN in skip_emulated_instruction to detect any attempt to
modify the guest's RIP during an SGX enclave VMEXIT.  Emulating an
instruction originating in an enclave is not possible as we don't have
access to the actual registers/memory related to the instruction, e.g.
the GPRs are loaded with synthetic state upon exiting the enclave.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>