KVM: x86: Define new #PF SGX error code bit
Page faults that are signaled by the SGX Enclave Page Cache Map (EPCM),
as opposed to the traditional IA32/EPT page tables, set an SGX bit in
the error code to indicate that the #PF was induced by SGX.  KVM will
need to emulate this behavior as part of its trap-and-execute scheme for
virtualizing SGX Launch Control and to support live migration, e.g. to
inject SGX-induced #PFs if EINIT faults in the host.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>