x86: intel_sgx: add launch control support in SGX core
KVM support for SGX, and more specifically SGX launch control, will
need to execute EINIT on behalf of its guests, writing the LE hash
MSRs as needed.  KVM will trap and execute EINIT instead of context
switching the MSRs because writing the MSRs is extraordinarily
expensive, the MSRs are not allowed in the hardware-accelerated
VMCS lists and EINIT is a relatively infrequent event, especially
compared to VMEntry/VMExit.

Move the management of the LE hash MSRs into SGX core and change
the optimization behavior to assume multiple writers.  Previously,
where the driver is the only writer, the most optimal approach is
to assume the MSRs hold the correct value, i.e. blindly execute
EINIT and only update the MSRs if EINIT fails.  With multiple
writers and usage models, e.g. driver and KVM, blindly executing
EINIT is more likely to fail.  To avoid unnecessary WRMSRs, add
a per-cpu cache of the MSR values.

Add an exported bool, sgx_lc_enabled, that is true iff SGX Launch
Control is supported and enabled in feature control MSR.  Add an
EINIT wrapper, sgx_einit, to handle setting the Launch Enclave
hash MSRs in conjunction with EINIT.  sgx_einit is a effecitvely
a pass-through to __einit when LC is disabled.  If LC is enabled,
disable pre-emption, and conditionally write the LE hash MSRs and
call __einit while preemption is disabled.  This needs to be done
for all enclaves regardless of whether or not they have a valid
EINIT token, as the LE hash value is always checked, either against
the token (its signer) or against the enclave itself.

Modify the SGX driver to use sgx_einit and remove the driver's own
code for updating the LE hash MSRs.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>