x86/sgx: Add new ENCLS leafs related to VMM EPC oversubscription
Add wrappers for new ENCLS leafs ERDINFO, ETRACKC, ELDBC and ELDUC.

ETRACKC and ELD*C are variants of existing leafs that return an error
code instead of generating a #GP if an EPC conflict is encountered.  The
"C" variants enable a VMM to swap its guests' pages in/out of the EPC
without generating faults in the VMM.

ERDINFO provides introspection capabilities on EPC pages and is
primarily intended for VMM use when evicting an EPC page, e.g. to
determine if the page being evicted should be reloaded in the BLOCKED or
UNBLOCKED state.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>