x86: intel_sgx: add new ENCLS leafs related to VMM EPC oversubscription
Define new ENCLS leafs ERDINFO, ETRACKC, ELDBC and ELDUC and add
their architectural collateral.  ETRACKC and ELD*C are variants of
existing functions that return an error code instead of generating
a #GP if an EPC conflict is encountered.  These variants enable a
VMM to swap its VM's pages in/out of the EPC without generating
faults in the VMM.  ERDINFO provides introspection capabilities on
EPC pages and is primarily intended for VMM use when evicting an
EPC page, e.g. to determine if the page being evicted should be
reloaded in the BLOCKED or UNBLOCKED state.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>