intel_sgx: use ioctl() for LE communication instead of file read/write
Replace the read/write mechanism of handling token requests with a
set of IOCTLs specific to the LE context.  This allows the userspace
helper to signal an error to the driver, e.g. if it failed to launch
its enclave.

If the userpsace helper reports an error, log it and return -EIO to
(actual) userspace.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>