x86: intel_sgx: add VA page swapping w/ EPC context encapsulation
Implement VA swapping using a tree structure not unlike x86 page
tables.  Encapsulate the VA page management in an "EPC context"
to allow the driver and KVM to share the VA management code.  An
EPC context represents the scope of a single VA page structure,
e.g. an enclave in the driver or a VM in KVM.

Do not convert the SGX driver to use the common VA swapping at
this time.  This temporarily results in a less-than-desirable
scenario of the driver and core both defining 'struct sgx_va_page',
but will not cause problems so long as the driver does not use
any of the new core features.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>