KVM: vmx: conditionally expose ENCLS_C leafs to guest
ENCLS_C leafs do not require any additional emulation, but because
ERDINFO provides access to the translated address of the SECS, i.e.
the HPA of the EPC when EPT is disabled, only allow the leafs to be
exposed to the guest when EPT is enabled (for now).

Alternatively, we could expose ERDINFO by intercepting ECREATE and
ELD* in order to stuff the appropriate GPA via ENCLV[ESETCONTEXT].
This approach may be introduced in a future patch, especially since
similar behavior is required to support EPC oversubscription without
EPT or when nesting is enabled (to translate L2 GPAs to L1 GPAs).

Intercepting and emulating ERDINFO to hide the HPA, i.e. by reverse
translating the HPA to GPA, isn't viable as there is no way to know
which VM loaded the EPC page (assuming L1 is running one or more L2
VMs, which is the primary use case for ERDINFO), nor could we ensure
the HPA to GPA rmap would be valid at the time of ERDINFO even if we
knew which rmap to use and could guarantee a 1:1 mapping of HPA:GPA.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>