KVM: VMX: Conditionally expose ENCLS_C leafs to guest
Advertise ENCLS_C functions to userspace/guest when EPT (TDP) is enabled
and the leafs are supported in the host.

ENCLS_C leafs do not require any additional emulation, but because
ERDINFO provides access to the translated address of the SECS, i.e. the
HPA of the EPC when EPT is disabled, only allow the leafs to be exposed
to the guest when EPT is enabled (for now).

Alternatively, ERDINFO could be exposed without EPT by intercepting
ECREATE and ELD* in order to stuff the appropriate GPA into the SECS via
ENCLV[ESETCONTEXT], but it is extremely unlikely there is a legit use
case for enabling nested EPC oversubscription with shadow paging.

Intercepting and emulating ERDINFO to hide the HPA, i.e. by reverse
translating the HPA to GPA, isn't viable as there is no way to know
which L2 guest loaded the EPC page if L1 is running one or more L2 VMs,
which is the primary use case for ERDINFO.  Nor could KVM ensure the HPA
to GPA rmap would be valid at the time of ERDINFO even if KVM knew which
rmap to use and could guarantee a 1:1 mapping of HPA:GPA.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>