intel_sgx: use a goto to handle dead enclave to reduce indentation
A future change to consolidate sgx_ewb and sgx_eld* logic into a common
function will double the number of parameters that need to be passed to
sgx_ewb().  Preemptively reduce the indentation of the sgx_ewb() call
sites by handling SGX_ENCL_DEAD via goto so that the upcoming change
does not result in absurd line usage due to wrapping parameters.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>