x86: intel_sgx: define ELDB, and add ECREATE and ELD* helpers
In the short term, the ECREATE and ELD* helpers will be used by
KVM when intercepting and executing their associated leafs.  In
the long term, the ELD*C leafs will also be used by KVM as part
of EPC oversubscription (to avoid #GPs when ELD* would conflict
with guest EPC activity).