KVM: vmx: add support for exposing ENCLS_C to VM without EPT
Because ECREATE and ELD* store the HPA of the SECS when EPT is
disabled, when ENCLS_C is exposed to the guest, i.e. the guest
can see SECS.enclavecontext via ERDINFO, we must intercept
ECREATE and ELD* in order to stuff SECS.enclavecontext with
L1's GPA in place of L0's HPAs.  Unfortunately, even though
we're only interested in ELD* of the SECS, we have to intercept
all instances of ELD* because that is the granularity we are
given.  Nor is it possible to prevent L1 (or L2) from evicting
an SECS without introducing unexpected behavior, e.g. we could
use EINCVIRTCHILD to prevent eviction, but that would cause
EWB to unexpectedly fail in the guest (and we can't trap on
failure because it's not a #GP condition).

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>